+++ R/driver-postgres.R
@@ -10,6 +10,7 @@ NULL
     time = "TIME",
     binary = "bytea",
     integer = "INTEGER",
+    int64 = "INTEGER",
     double = "DOUBLE PRECISION",
     character = "TEXT",
     logical = "BOOLEAN",