+++ R/driver-bigquery.R
@@ -10,6 +10,7 @@ NULL
     date = "DATE",
     binary = "BYTES",
     integer = "INT64",
+    int64 = "INT64",
     double = "FLOAT64",
     character = "STRING",
     logical = "BOOL",