+++ R/odbc-data-type.R
@@ -62,6 +62,7 @@ odbcDataType.default <- function(con, obj, ...) {
     time = "TIME",
     binary = "VARBINARY(255)",
     integer = "INTEGER",
+    int64 = "INTEGER",
     double = "DOUBLE PRECISION",
     character = "VARCHAR(255)",
     logical = "BIT", # only valid if DB supports Null fields
@@ -91,6 +92,9 @@ object_type <- function(obj) {
   }
   if (is(obj, "difftime")) {
     return("time")
+  }
+  if (is(obj, "integer64")) {
+    return("int64")
   }
 
   return(typeof(obj))