+++ R/utils.R
@@ -112,3 +112,9 @@ escapePattern <- function(x, charsToEsc = c("_"), escChar = "\\\\") {
   x <- gsub(pattern, replace, x)
   I(x)
 }
+
+is_windows <- function() {
+  identical(.Platform$OS.type, "windows")
+}
+
+compact <- function(x) x[!vapply(x, is.null, logical(1))]