+++ NEWS.md
@@ -2,10 +2,15 @@
 
 * `databricks()` now works with manually supplied `pwd` and `uid` (#690).
 
+* SQL Server: correctly enumerate schemas across databases in connections pane
+  (@detule, #527).
+
 * Added a function `odbcListConfig()` to help locate configuration files on
   macOS and Linux (@simonpcouch, #565).
 
 * Use correct parent class for Oracle (#685).
+
+* Removed the `driver_info` class from `dbGetInfo()` output (@simonpcouch, #667).
 
 # odbc 1.4.0