+++ NEWS.md
@@ -1,5 +1,8 @@
 # odbc (development version)
 
+* The Microsoft SQL Server method for `odbcDataType()` will now return `"BIGINT"`
+  for integer64 objects rather than converting to `"FLOAT"` (@simonpcouch, #698).
+
 * `databricks()` now works with manually supplied `pwd` and `uid` (#690).
 
 * SQL Server: correctly enumerate schemas across databases in connections pane