+++ NEWS.md
@@ -1,8 +1,21 @@
 # odbc (development version)
 
+* The Microsoft SQL Server method for `odbcDataType()` will now return `"BIGINT"`
+  for integer64 objects rather than converting to `"FLOAT"` (@simonpcouch, #698).
+
+* `databricks()` now works with manually supplied `pwd` and `uid` (#690).
+
+* SQL Server: correctly enumerate schemas across databases in connections pane
+  (@detule, #527).
+
+* Added a function `odbcListConfig()` to help locate configuration files on
+  macOS and Linux (@simonpcouch, #565).
+
 * Use correct parent class for Oracle (#685).
 
 * dbAppendTable: Improve performance by checking existence once (#691).
+
+* Removed the `driver_info` class from `dbGetInfo()` output (@simonpcouch, #667).
 
 # odbc 1.4.0