+++ NEWS.md
@@ -1,4 +1,4 @@
-# odbc (development version)
+# odbc 1.4.1
 
 * New `odbcListConfig()` lists configuration files on Mac and Linux 
   (@simonpcouch, #565).